requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
SQLAlchemy==2.0.9
alembic==1.10.3
asyncpg==0.27.0
aiohttp==3.8.3
Pillow==9.0.1
psycopg2-binary==2.9.6